+++ src/DependencyInjection/HeimrichHannotContaoNewsAlertExtension.php
@@ -33,6 +33,7 @@ class HeimrichHannotContaoNewsAlertExtension extends Extension
             $container, new FileLocator(__DIR__ . '/../Resources/config')
         );
         $loader->load('services.yml');
+        $loader->load('commands.yml');
     }
 
 }